modify
git-svn-id: https://svn.baidu.com/app/search/sep/trunk/anti-spam/gezi@7461 fa64baa9-71d1-4fed-97ae-c15534abce97

+++ include/Numeric/Vector/Vector.h
@@ -73,7 +73,7 @@ namespace gezi {
 				{
 					for (string val_ : inputs)
 					{
-						Add(val_);
+						Add(DOUBLE(val_));
 					}
 				}
 			}
